text2video/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-multipart==0.0.9

celery==5.4.0
redis==5.0.8
pyttsx3==2.90

Pillow>=10.0.0
numpy>=1.24.0
